+++ CHANGELOG.md
@@ -68,6 +68,7 @@
 * Review existing FOLIO user records on duplicates view. Refs UIU-3224.
 * Create new permission 'Users: Can merge patron preregistration data'. Refs UIU-3237.
 * Confirmation dialog when JIT AuthUser creation is required. Refs UIU-3192.
+*  "borrower.patronGroup" print patron group's name based on id. Refs UIU-3228.
 * Create new user record from preregistration data when no matching folio user is found. Refs UIU-3223.
 
 ## [10.1.2](https://github.com/folio-org/ui-users/tree/v10.1.2) (2024-09-05)